requirements.txt
Athena_v1/requirements.txt

pip install -r requirements.txt 로 설치합니다.

fastapi
uvicorn
websockets
python-dotenv
pyupbit
pandas
numpy
SQLAlchemy
aiohttp
asyncio
requests